+++ services/workflows-service/scripts/alerts/generate-alerts.ts
@@ -1,9 +1,21 @@
+import {
+  ALERT_DEDUPE_STRATEGY_DEFAULT,
+  daysToMinutesConverter,
+  MerchantAlertLabel,
+  SEVEN_DAYS,
+  THREE_DAYS,
+  TWENTY_ONE_DAYS,
+} from '@/alert/consts';
+import { TDedupeStrategy } from '@/alert/types';
+import { AggregateType, TIME_UNITS } from '@/data-analytics/consts';
 import { InlineRule } from '@/data-analytics/types';
+import { InputJsonValue, PrismaTransaction } from '@/types';
+import { faker } from '@faker-js/faker';
 import {
   AlertSeverity,
   AlertState,
   AlertStatus,
-  AlertType,
+  MonitoringType,
   PaymentMethod,
   Prisma,
   PrismaClient,
@@ -11,9 +23,6 @@ import {
   TransactionDirection,
   TransactionRecordType,
 } from '@prisma/client';
-import { faker } from '@faker-js/faker';
-import { AggregateType, TIME_UNITS } from '@/data-analytics/consts';
-import { InputJsonValue, PrismaTransaction } from '@/types';
 
 const tags = [
   ...new Set([
@@ -45,7 +54,8 @@ export const ALERT_DEFINITIONS = {
         },
         paymentMethods: [PaymentMethod.credit_card],
         excludePaymentMethods: false,
-        timeAmount: 7,
+
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
         amountThreshold: 1000,
         groupBy: ['counterpartyBeneficiaryId'],
@@ -74,7 +84,7 @@ export const ALERT_DEFINITIONS = {
         paymentMethods: [PaymentMethod.credit_card],
         excludePaymentMethods: true,
 
-        timeAmount: 7,
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
 
         amountThreshold: 1000,
@@ -105,7 +115,7 @@ export const ALERT_DEFINITIONS = {
         paymentMethods: [PaymentMethod.credit_card],
         excludePaymentMethods: false,
 
-        timeAmount: 7,
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
 
         amountThreshold: 5,
@@ -133,7 +143,8 @@ export const ALERT_DEFINITIONS = {
         },
         paymentMethods: [PaymentMethod.credit_card],
         excludePaymentMethods: true,
-        timeAmount: 7,
+
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
         amountBetween: { min: 500, max: 999 },
         amountThreshold: 5,
@@ -141,6 +152,7 @@ export const ALERT_DEFINITIONS = {
     },
   },
   HCAI_CC: {
+    enabled: false,
     defaultSeverity: AlertSeverity.medium,
     description:
       'High Cumulative Amount - Total sum of inbound credit card transactions received from counterparty is greater than a limit over a set period of time',
@@ -161,7 +173,7 @@ export const ALERT_DEFINITIONS = {
         paymentMethods: [PaymentMethod.credit_card],
         excludePaymentMethods: false,
 
-        timeAmount: 7,
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
 
         amountThreshold: 3000,
@@ -169,6 +181,7 @@ export const ALERT_DEFINITIONS = {
     },
   },
   HACI_APM: {
+    enabled: false,
     defaultSeverity: AlertSeverity.medium,
     description:
       'High Cumulative Amount - Total sum of inbound non-traditional payment transactions received from counterparty is greater than a limit over a set period of time',
@@ -189,7 +202,7 @@ export const ALERT_DEFINITIONS = {
         paymentMethods: [PaymentMethod.credit_card],
         excludePaymentMethods: true,
 
-        timeAmount: 7,
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
 
         amountThreshold: 3000,
@@ -197,6 +210,7 @@ export const ALERT_DEFINITIONS = {
     },
   },
   HVIC_CC: {
+    enabled: false,
     defaultSeverity: AlertSeverity.medium,
     description:
       'High Velocity - High number of inbound credit card transactions received from a Counterparty over a set period of time',
@@ -217,7 +231,7 @@ export const ALERT_DEFINITIONS = {
         paymentMethods: [PaymentMethod.credit_card],
         excludePaymentMethods: false,
 
-        timeAmount: 7,
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
 
         amountThreshold: 2,
@@ -225,6 +239,7 @@ export const ALERT_DEFINITIONS = {
     },
   },
   HVIC_APM: {
+    enabled: false,
     defaultSeverity: AlertSeverity.medium,
     description:
       'High Velocity - High number of inbound non-traditional payment transactions received from a Counterparty over a set period of time',
@@ -245,7 +260,7 @@ export const ALERT_DEFINITIONS = {
         paymentMethods: [PaymentMethod.credit_card],
         excludePaymentMethods: true,
 
-        timeAmount: 7,
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
 
         amountThreshold: 2,
@@ -265,8 +280,10 @@ export const ALERT_DEFINITIONS = {
         transactionType: [TransactionRecordType.chargeback],
         paymentMethods: [PaymentMethod.credit_card],
         amountThreshold: 14,
-        timeAmount: 7,
+
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
+
         groupBy: ['counterpartyOriginatorId'],
         havingAggregate: AggregateType.COUNT,
       },
@@ -284,9 +301,12 @@ export const ALERT_DEFINITIONS = {
       options: {
         transactionType: [TransactionRecordType.chargeback],
         paymentMethods: [PaymentMethod.credit_card],
+
         amountThreshold: 5_000,
-        timeAmount: 7,
+
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
+
         groupBy: ['counterpartyOriginatorId'],
         havingAggregate: AggregateType.SUM,
       },
@@ -303,9 +323,12 @@ export const ALERT_DEFINITIONS = {
       options: {
         transactionType: [TransactionRecordType.refund],
         paymentMethods: [PaymentMethod.credit_card],
+
         amountThreshold: 14,
-        timeAmount: 7,
+
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
+
         groupBy: ['counterpartyOriginatorId'],
         havingAggregate: AggregateType.COUNT,
       },
@@ -323,7 +346,8 @@ export const ALERT_DEFINITIONS = {
         transactionType: [TransactionRecordType.refund],
         paymentMethods: [PaymentMethod.credit_card],
         amountThreshold: 5_000,
-        timeAmount: 7,
+
+        timeAmount: SEVEN_DAYS,
         timeUnit: TIME_UNITS.days,
         groupBy: ['counterpartyOriginatorId'],
         havingAggregate: AggregateType.SUM,
@@ -335,6 +359,10 @@ export const ALERT_DEFINITIONS = {
     defaultSeverity: AlertSeverity.high,
     description:
       'High Percentage of Chargebacks - High percentage of chargebacks over a set period of time',
+    dedupeStrategy: {
+      mute: false,
+      cooldownTimeframeInMinutes: daysToMinutesConverter(TWENTY_ONE_DAYS),
+    },
     inlineRule: {
       id: 'HPC',
       fnName: 'evaluateHighTransactionTypePercentage',
@@ -344,18 +372,18 @@ export const ALERT_DEFINITIONS = {
         subjectColumn: 'counterpartyOriginatorId',
         minimumCount: 3,
         minimumPercentage: 50,
-        timeAmount: 21,
+        timeAmount: TWENTY_ONE_DAYS,
         timeUnit: TIME_UNITS.days,
       },
     },
   },
   TLHAICC: {
-    enabled: true,
+    enabled: false,
     defaultSeverity: AlertSeverity.medium,
     description: `Transaction Limit - Historic Average - Inbound - Inbound transaction exceeds client's historical average`,
     inlineRule: {
       id: 'TLHAICC',
-      fnName: 'evaluateTransactionLimitHistoricAverageInbound',
+      fnName: 'evaluateTransactionAvg',
       subjects: ['counterpartyId'],
       options: {
         transactionDirection: TransactionDirection.inbound,
@@ -366,16 +394,19 @@ export const ALERT_DEFINITIONS = {
         },
         minimumTransactionAmount: 100,
         transactionFactor: 1,
+        customerType: undefined,
+        timeUnit: undefined,
+        timeAmount: undefined,
       },
     },
   },
   TLHAIAPM: {
-    enabled: true,
+    enabled: false,
     defaultSeverity: AlertSeverity.medium,
     description: `Transaction Limit - Historic Average - Inbound - Inbound transaction exceeds client's historical average`,
     inlineRule: {
       id: 'TLHAIAPM',
-      fnName: 'evaluateTransactionLimitHistoricAverageInbound',
+      fnName: 'evaluateTransactionAvg',
       subjects: ['counterpartyId'],
       options: {
         transactionDirection: TransactionDirection.inbound,
@@ -386,94 +417,367 @@ export const ALERT_DEFINITIONS = {
         },
         minimumTransactionAmount: 100,
         transactionFactor: 1,
+        customerType: undefined,
+        timeUnit: undefined,
+        timeAmount: undefined,
       },
     },
   },
-} as const satisfies Record<
-  string,
-  {
-    inlineRule: InlineRule;
-    defaultSeverity: AlertSeverity;
-    enabled?: boolean;
-    description?: string;
-  }
+  PGAICT: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.medium,
+    description: `An Credit card inbound transaction value was over the peer group average within a set period of time`,
+    inlineRule: {
+      id: 'PGAICT',
+      fnName: 'evaluateTransactionAvg',
+      subjects: ['counterpartyId'],
+      options: {
+        transactionDirection: TransactionDirection.inbound,
+        minimumCount: 2,
+        paymentMethod: {
+          value: PaymentMethod.credit_card,
+          operator: '=',
+        },
+        minimumTransactionAmount: 100,
+        transactionFactor: 2,
+
+        customerType: 'test',
+        timeAmount: SEVEN_DAYS,
+        timeUnit: TIME_UNITS.days,
+      },
+    },
+  },
+  PGAIAPM: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.medium,
+    description: `An non credit card inbound transaction value was over the peer group average within a set period of time`,
+    inlineRule: {
+      id: 'PGAIAPM',
+      fnName: 'evaluateTransactionAvg',
+      subjects: ['counterpartyId'],
+      options: {
+        transactionDirection: TransactionDirection.inbound,
+        minimumCount: 2,
+        paymentMethod: {
+          value: PaymentMethod.credit_card,
+          operator: '!=',
+        },
+        customerType: 'test',
+        minimumTransactionAmount: 100,
+        transactionFactor: 2,
+
+        timeAmount: SEVEN_DAYS,
+        timeUnit: TIME_UNITS.days,
+      },
+    },
+  },
+  DORMANT: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.high,
+    description: `First activity of client after a long period of dormancy`,
+    inlineRule: {
+      id: 'DORMANT',
+      fnName: 'evaluateDormantAccount',
+      subjects: ['counterpartyId'],
+      options: {
+        timeAmount: 180,
+        timeUnit: TIME_UNITS.days,
+      },
+    },
+  },
+  HVHAI_CC: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.medium,
+    description: `Total number of incoming credit cards transactions exceeds client’s historical average`,
+    inlineRule: {
+      id: 'HVHAI_CC',
+      fnName: 'evaluateHighVelocityHistoricAverage',
+      subjects: ['counterpartyId'],
+      options: {
+        transactionDirection: TransactionDirection.inbound,
+        minimumCount: 3,
+        transactionFactor: 2,
+        paymentMethod: {
+          value: PaymentMethod.credit_card,
+          operator: '=',
+        },
+        activeUserPeriod: {
+          timeAmount: 180,
+        },
+        lastDaysPeriod: {
+          timeAmount: THREE_DAYS,
+        },
+        timeUnit: TIME_UNITS.days,
+      },
+    },
+  },
+  HVHAI_APM: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.medium,
+    description: `Total number of incoming credit cards transactions exceeds client’s historical average`,
+    inlineRule: {
+      id: 'HVHAI_APM',
+      fnName: 'evaluateHighVelocityHistoricAverage',
+      subjects: ['counterpartyId'],
+      options: {
+        transactionDirection: TransactionDirection.inbound,
+        minimumCount: 3,
+        transactionFactor: 2,
+        paymentMethod: {
+          value: PaymentMethod.credit_card,
+          operator: '!=',
+        },
+        activeUserPeriod: {
+          timeAmount: 180,
+        },
+        lastDaysPeriod: {
+          timeAmount: THREE_DAYS,
+        },
+        timeUnit: TIME_UNITS.days,
+      },
+    },
+  },
+  MMOC_CC: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.high,
+    description: `Card numbers that are appearing in too many different merchant IDs for credit card transactions`,
+    inlineRule: {
+      id: 'MMOC_CC',
+      fnName: 'evaluateMultipleMerchantsOneCounterparty',
+      subjects: ['counterpartyId'],
+      options: {
+        excludedCounterparty: {
+          counterpartyBeneficiaryIds: ['9999999999999999', '999999______9999'],
+          counterpartyOriginatorIds: [],
+        },
+        minimumCount: 2,
+        timeAmount: SEVEN_DAYS,
+        timeUnit: TIME_UNITS.days,
+      },
+    },
+  },
+  MMOC_APM: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.high,
+    description: `Card numbers that are appearing in too many different merchant IDs for non credit card transactions`,
+    inlineRule: {
+      id: 'MMOC_APM',
+      fnName: 'evaluateMultipleMerchantsOneCounterparty',
+      subjects: ['counterpartyId'],
+      options: {
+        excludedCounterparty: {
+          counterpartyBeneficiaryIds: ['9999999999999999', '999999______9999'],
+          counterpartyOriginatorIds: [],
+        },
+        minimumCount: 2,
+        timeAmount: SEVEN_DAYS,
+        timeUnit: TIME_UNITS.days,
+      },
+    },
+  },
+} as const satisfies Record<string, Parameters<typeof getAlertDefinitionCreateData>[0]>;
+
+export const MERCHANT_MONITORING_ALERT_DEFINITIONS = {
+  MERCHANT_ONGOING_RISK_ALERT_RISK_INCREASE: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.low,
+    monitoringType: MonitoringType.ongoing_merchant_monitoring,
+    description: 'Monitor ongoing risk changes',
+    inlineRule: {
+      id: 'MERCHANT_ONGOING_RISK_ALERT_RISK_INCREASE',
+      fnName: 'checkMerchantOngoingAlert',
+      subjects: ['businessId', 'projectId'],
+      options: {
+        increaseRiskScore: 20,
+      },
+    },
+  },
+  MERCHANT_ONGOING_RISK_ALERT_THRESHOLD: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.high,
+    monitoringType: MonitoringType.ongoing_merchant_monitoring,
+    description: 'Monitor ongoing risk changes',
+    inlineRule: {
+      id: 'MERCHANT_ONGOING_RISK_ALERT_THRESHOLD',
+      fnName: 'checkMerchantOngoingAlert',
+      subjects: ['businessId', 'projectId'],
+      options: {
+        maxRiskScoreThreshold: 60,
+      },
+    },
+  },
+  MERCHANT_ONGOING_RISK_ALERT_PERCENTAGE: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.medium,
+    monitoringType: MonitoringType.ongoing_merchant_monitoring,
+    description: 'Monitor ongoing risk changes',
+    inlineRule: {
+      id: 'MERCHANT_ONGOING_RISK_ALERT_PERCENTAGE',
+      fnName: 'checkMerchantOngoingAlert',
+      subjects: ['businessId', 'projectId'],
+      options: {
+        increaseRiskScorePercentage: 30,
+      },
+    },
+  },
+} as const satisfies Partial<
+  Record<
+    keyof typeof MerchantAlertLabel | string,
+    {
+      inlineRule: InlineRule & InputJsonValue;
+      monitoringType: MonitoringType;
+      defaultSeverity: AlertSeverity;
+      enabled?: boolean;
+      description?: string;
+    }
+  >
 >;
 
 export const getAlertDefinitionCreateData = (
   {
     inlineRule,
     defaultSeverity,
-    label,
     description,
+    monitoringType = MonitoringType.transaction_monitoring,
     enabled = false,
+    dedupeStrategy = ALERT_DEDUPE_STRATEGY_DEFAULT,
+    correlationId,
   }: {
-    label: string;
     inlineRule: InlineRule;
     defaultSeverity: AlertSeverity;
-    enabled?: boolean;
+    monitoringType?: MonitoringType;
+    enabled: boolean;
+    dedupeStrategy?: Partial<TDedupeStrategy>;
     description?: string;
+    correlationId?: string;
   },
   project: Project,
   createdBy: string = 'SYSTEM',
-) => ({
-  label: label,
-  type: faker.helpers.arrayElement(Object.values(AlertType)) as AlertType,
-  name: inlineRule.id,
-  enabled: enabled ?? false,
-  description: description || faker.lorem.sentence(),
-  rulesetId: `set-${inlineRule.id}`,
-  defaultSeverity,
-  ruleId: inlineRule.id,
-  createdBy: createdBy,
-  modifiedBy: createdBy,
-  dedupeStrategy: {
-    mute: false,
-    cooldownTimeframeInMinutes: faker.datatype.number({ min: 60, max: 3600 }),
+  extraColumns: {
+    crossEnvKey: string;
   },
-  config: { config: {} },
-  inlineRule,
-  tags: [faker.helpers.arrayElement(tags), faker.helpers.arrayElement(tags)],
-  additionalInfo: {},
-  projectId: project.id,
-});
+) => {
+  const id = inlineRule.id;
+
+  return {
+    enabled,
+    defaultSeverity,
+    dedupeStrategy: {
+      ...ALERT_DEDUPE_STRATEGY_DEFAULT,
+      ...(dedupeStrategy ?? {}),
+    },
+    monitoringType: monitoringType ?? MonitoringType.transaction_monitoring,
+    inlineRule,
+    correlationId: correlationId || id,
+    name: id,
+    rulesetId: `set-${id}`,
+    description: description,
+    ruleId: id,
+    createdBy: createdBy,
+    modifiedBy: createdBy,
+    config: { config: {} },
+    tags: [],
+    additionalInfo: {},
+    crossEnvKey: extraColumns.crossEnvKey,
+    project: {
+      connect: {
+        id: project.id,
+      },
+    },
+  } as Prisma.AlertDefinitionCreateInput;
+};
 
 export const generateAlertDefinitions = async (
   prisma: PrismaClient | PrismaTransaction,
   {
-    createdBy = 'SYSTEM',
     project,
+    createdBy = 'SYSTEM',
+    alertsDef = ALERT_DEFINITIONS,
   }: {
+    alertsDefConfiguration?:
+      | typeof ALERT_DEFINITIONS
+      | typeof MERCHANT_MONITORING_ALERT_DEFINITIONS;
     createdBy?: string;
     project: Project;
+    alertsDef?: Partial<typeof ALERT_DEFINITIONS>;
   },
+  {
+    crossEnvKeyPrefix = undefined,
+  }: {
+    crossEnvKeyPrefix?: string;
+  } = {},
 ) =>
   Promise.all(
-    Object.entries(ALERT_DEFINITIONS)
-      .filter(([_, alert]) => 'enabled' in alert && alert.enabled)
-      .map(([label, data]) =>
-        prisma.alertDefinition.upsert({
-          where: { label_projectId: { label: label, projectId: project.id } },
-          create: getAlertDefinitionCreateData({ label, ...data }, project, createdBy),
-          update: getAlertDefinitionCreateData({ label, ...data }, project, createdBy),
-          include: {
-            alert: true,
-          },
-        }),
-      ),
+    Object.values(alertsDef)
+      .map(alert => ({
+        correlationId: alert.inlineRule.id,
+        ...alert,
+      }))
+      .filter(alert => alert.enabled)
+      .map(async alertDef => {
+        const extraColumns = {
+          crossEnvKey: crossEnvKeyPrefix
+            ? `${crossEnvKeyPrefix}_${alertDef.inlineRule.id}`
+            : project.name
+                .toUpperCase()
+                .replace(' ', '_')
+                .replace(/[_-]?PROJECT[_-]?/g, 'P')
+                .replace(/[_-]?DEFAULT[_-]?/g, '')
+                .replace('-', '_')
+                .replace('__', '_') +
+              '_' +
+              alertDef.inlineRule.id,
+        };
+        let dbRes;
+        try {
+          dbRes = await prisma.alertDefinition.upsert({
+            where: {
+              crossEnvKey: extraColumns.crossEnvKey,
+            },
+            create: getAlertDefinitionCreateData(alertDef, project, createdBy, extraColumns),
+            update: getAlertDefinitionCreateData(alertDef, project, createdBy, extraColumns),
+            include: {
+              alert: true,
+            },
+          });
+        } catch (error) {
+          console.error(error);
+          console.error('Error creating alert definition', alertDef, extraColumns);
+          throw error;
+        }
+
+        return dbRes;
+      }),
   );
 
-const generateFakeAlert = ({
-  severity,
-  counterparyIds,
-  agentUserIds,
-}: {
-  severity: AlertSeverity;
-  counterparyIds: string[];
-  agentUserIds: string[];
-}): Omit<Prisma.AlertCreateManyAlertDefinitionInput, 'projectId'> => {
-  const counterypartyId = faker.helpers.arrayElement(
-    counterparyIds.map(id => ({ counterpartyId: id })),
-  );
+const generateFakeAlert = (
+  options: {
+    severity: AlertSeverity;
+    agentUserIds: string[];
+  } & (
+    | {
+        counterpartyIds: string[];
+      }
+    | {
+        businessIds: string[];
+      }
+  ),
+): Omit<Prisma.AlertCreateManyAlertDefinitionInput, 'projectId'> => {
+  const { severity, agentUserIds } = options;
+
+  let businessId: { businessId: string } | {} = {};
+  let counterpartyId: { counterpartyId: string } | {} = {};
+
+  if ('businessIds' in options) {
+    // For merchant monitoring alerts
+    businessId = faker.helpers.arrayElement(options.businessIds.map(id => ({ businessId: id })));
+  } else if ('counterpartyIds' in options) {
+    // For transaction alerts
+    counterpartyId = faker.helpers.arrayElement(
+      options.counterpartyIds.map(id => ({ counterpartyId: id })),
+    );
+  }
+
   // In chance of 1 to 5, assign an agent to the alert
   const assigneeId =
     faker.datatype.number({ min: 1, max: 5 }) === 1
@@ -493,29 +797,39 @@ const generateFakeAlert = ({
     } as InputJsonValue,
     severity,
     assigneeId,
-    ...counterypartyId,
+    ...counterpartyId,
+    ...businessId,
   };
 };
 
-export const generateFakeAlertsAndDefinitions = async (
+export const seedTransactionsAlerts = async (
   prisma: PrismaClient,
   {
     project,
-    counterparyIds,
+    businessIds,
+    counterpartyIds,
     agentUserIds,
   }: {
     project: Project;
-    counterparyIds: string[];
+    businessIds: string[];
+    counterpartyIds: string[];
     agentUserIds: string[];
   },
 ) => {
-  const alertDefinitions = await generateAlertDefinitions(prisma, {
+  const transactionsAlertDef = await generateAlertDefinitions(prisma, {
+    alertsDefConfiguration: ALERT_DEFINITIONS,
     project,
     createdBy: faker.internet.userName(),
   });
 
-  await Promise.all(
-    alertDefinitions.map(alertDefinition =>
+  const merchantMonitoringAlertDef = await generateAlertDefinitions(prisma, {
+    alertsDefConfiguration: MERCHANT_MONITORING_ALERT_DEFINITIONS,
+    project,
+    createdBy: faker.internet.userName(),
+  });
+
+  await Promise.all([
+    ...transactionsAlertDef.map(alertDefinition =>
       prisma.alert.createMany({
         data: Array.from(
           {
@@ -525,7 +839,7 @@ export const generateFakeAlertsAndDefinitions = async (
             alertDefinitionId: alertDefinition.id,
             projectId: project.id,
             ...generateFakeAlert({
-              counterparyIds,
+              counterpartyIds,
               agentUserIds,
               severity: faker.helpers.arrayElement(Object.values(AlertSeverity)),
             }),
@@ -534,5 +848,5 @@ export const generateFakeAlertsAndDefinitions = async (
         skipDuplicates: true,
       }),
     ),
-  );
+  ]);
 };