+++ services/workflows-service/src/business-report/dto/get-business-reports.dto.ts
@@ -1,0 +1,19 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { BusinessReportType } from '@prisma/client';
+import { IsIn, IsNotEmpty, IsOptional, IsString } from 'class-validator';
+
+export class GetBusinessReportsDto {
+  @ApiProperty({
+    required: true,
+  })
+  @IsNotEmpty()
+  @IsString()
+  businessId!: string;
+
+  @ApiProperty({
+    required: false,
+  })
+  @IsIn(Object.values(BusinessReportType))
+  @IsOptional()
+  type?: BusinessReportType;
+}