+++ services/workflows-service/src/workflow/hook-callback-handler.service.ts
@@ -13,6 +13,18 @@ import { BusinessReportType, Customer, WorkflowRuntimeData } from '@prisma/clien
 import fs from 'fs';
 import { get, isObject, set } from 'lodash';
 import * as tmp from 'tmp';
+import { AlertService } from '@/alert/alert.service';
+import { EndUserService } from '@/end-user/end-user.service';
+import { z } from 'zod';
+import { EndUserActiveMonitoringsSchema } from '@/end-user/end-user.schema';
+
+const ReportWithRiskScoreSchema = z
+  .object({
+    summary: z.object({
+      riskScore: z.number(),
+    }),
+  })
+  .passthrough();
 
 @Injectable()
 export class HookCallbackHandlerService {
@@ -21,7 +33,9 @@ export class HookCallbackHandlerService {
     protected readonly customerService: CustomerService,
     protected readonly businessReportService: BusinessReportService,
     protected readonly businessService: BusinessService,
+    protected readonly alertService: AlertService,
     private readonly logger: AppLoggerService,
+    private readonly endUserService: EndUserService,
   ) {}
 
   async handleHookResponse({
@@ -39,12 +53,26 @@ export class HookCallbackHandlerService {
     currentProjectId: TProjectId;
   }) {
     if (processName === 'kyc-unified-api') {
-      return await this.mapCallbackDataToIndividual(
+      const context = await this.mapCallbackDataToIndividual(
         data,
         workflowRuntime,
         resultDestinationPath,
         currentProjectId,
       );
+
+      const aml = data.aml as { endUserId: string; hits: unknown[] } | undefined;
+
+      if (aml) {
+        await this.updateEndUserWithAmlData({
+          sessionId: data.id as string,
+          amlHits: aml.hits,
+          withActiveMonitoring: workflowRuntime.config.hasUboOngoingMonitoring ?? false,
+          endUserId: aml.endUserId,
+          projectId: currentProjectId,
+        });
+      }
+
+      return context;
     }
 
     if (processName === 'website-monitoring') {
@@ -58,7 +86,13 @@ export class HookCallbackHandlerService {
 
     if (processName === 'merchant-audit-report') {
       return await this.prepareMerchantAuditReportContext(
-        data,
+        data as {
+          reportData: Record<string, unknown>;
+          base64Pdf: string;
+          reportId: string;
+          reportType: string;
+          comparedToReportId?: string;
+        },
         workflowRuntime,
         resultDestinationPath,
         currentProjectId,
@@ -95,7 +129,8 @@ export class HookCallbackHandlerService {
     const customer = await this.customerService.getByProjectId(currentProjectId);
 
     const { context } = workflowRuntime;
-    const { reportData, base64Pdf, reportId, reportType } = data;
+    const { reportData: unvalidatedReportData, base64Pdf, reportId, reportType } = data;
+    const reportData = ReportWithRiskScoreSchema;
 
     const { documents, pdfReportBallerineFileId } =
       await this.__peristPDFReportDocumentWithWorkflowDocuments({
@@ -105,43 +140,112 @@ export class HookCallbackHandlerService {
         base64PDFString: base64Pdf as string,
       });
 
+    const reportRiskScore =
+      ReportWithRiskScoreSchema.parse(unvalidatedReportData).summary.riskScore;
+
     const business = await this.businessService.getByCorrelationId(context.entity.id, [
       currentProjectId,
     ]);
 
     if (!business) throw new BadRequestException('Business not found.');
 
-    await this.businessReportService.create({
-      data: {
-        type: reportType as BusinessReportType,
-        report: {
-          reportFileId: pdfReportBallerineFileId,
-          data: reportData as InputJsonValue,
-          reportId: reportId as string,
+    const currentReportId = reportId as string;
+    const existantBusinessReport = await this.businessReportService.findFirstOrThrow(
+      {
+        where: {
+          businessId: business.id,
+          reportId: currentReportId,
         },
-        businessId: business.id,
-        projectId: currentProjectId,
       },
-    });
+      [currentProjectId],
+    );
+
+    const businessReport = await this.businessReportService.upsert(
+      {
+        create: {
+          type: reportType as BusinessReportType,
+          riskScore: reportRiskScore as number,
+          report: {
+            reportFileId: pdfReportBallerineFileId,
+            data: reportData as InputJsonValue,
+          },
+          reportId: currentReportId,
+          businessId: business.id,
+          projectId: currentProjectId,
+        },
+        update: {
+          type: reportType as BusinessReportType,
+          riskScore: reportRiskScore,
+          report: {
+            reportFileId: pdfReportBallerineFileId,
+            data: reportData as InputJsonValue,
+          },
+        },
+        where: {
+          id: existantBusinessReport?.id,
+        },
+      },
+      [currentProjectId],
+    );
 
     set(workflowRuntime.context, resultDestinationPath, { reportData });
     workflowRuntime.context.documents = documents;
+
+    this.alertService
+      .checkOngoingMonitoringAlert(businessReport, business.companyName)
+      .then(() => {
+        this.logger.debug(`Alert Tested for ${currentReportId}}`);
+      })
+      .catch(error => {
+        this.logger.error(error);
+      });
 
     return context;
   }
 
   async prepareMerchantAuditReportContext(
-    data: AnyRecord,
+    data: Record<string, unknown>,
     workflowRuntime: WorkflowRuntimeData,
     resultDestinationPath: string,
     currentProjectId: TProjectId,
   ) {
-    const { reportData, base64Pdf, reportId, reportType } = data;
+    const { reportData, base64Pdf, reportId, reportType, comparedToReportId } = z
+      .object({
+        reportData: ReportWithRiskScoreSchema,
+        base64Pdf: z.string(),
+        reportId: z.string(),
+        reportType: z.string(),
+        comparedToReportId: z.string().optional(),
+      })
+      .parse(data);
+
     const { context } = workflowRuntime;
 
     const businessId = context.entity.id as string;
 
     const customer = await this.customerService.getByProjectId(currentProjectId);
+
+    if (comparedToReportId) {
+      const comparedToReport = await this.businessReportService.findFirstOrThrow(
+        {
+          where: {
+            businessId,
+            reportId: comparedToReportId,
+          },
+        },
+        [currentProjectId],
+      );
+
+      if (!comparedToReport) {
+        throw new BadRequestException('Compared to report not found.');
+      }
+
+      reportData.previousReport = {
+        summary: (comparedToReport.report as { data: { summary: { summary: unknown } } }).data
+          .summary,
+        reportType: comparedToReport.type,
+      };
+    }
 
     const { pdfReportBallerineFileId } = await this.__peristPDFReportDocumentWithWorkflowDocuments({
       context,
@@ -156,12 +260,16 @@ export class HookCallbackHandlerService {
       reportId,
     } as Record<string, object | string>;
 
+    const reportRiskScore = reportData.summary.riskScore;
+
     await this.businessReportService.create({
       data: {
         type: reportType as BusinessReportType,
         report: reportContent,
         businessId: businessId,
+        reportId: reportId as string,
         projectId: currentProjectId,
+        riskScore: reportRiskScore,
       },
     });
 
@@ -269,7 +377,7 @@ export class HookCallbackHandlerService {
     const customer = await this.customerService.getByProjectId(currentProjectId);
     const persistedDocuments = await this.workflowService.copyDocumentsPagesFilesAndCreate(
       documents as TDocumentsWithoutPageType,
-      // @ts-expect-error - we don't validate `context` is an object1
+      // @ts-expect-error - we don't validate `context` is an object
       context.entity.id || context.entity.ballerineEntityId,
       currentProjectId,
       customer.name,
@@ -378,7 +486,7 @@ export class HookCallbackHandlerService {
     const documentImages: AnyRecord[] = [];
 
     for (const image of data.images as Array<{ context?: string; content: string }>) {
-      const tmpFile = tmp.fileSync().name;
+      const tmpFile = tmp.fileSync({ keep: false }).name;
       const base64ImageContent = image.content.split(',')[1];
       const buffer = Buffer.from(base64ImageContent as string, 'base64');
       const fileType = await getFileMetadata({
@@ -430,5 +538,42 @@ export class HookCallbackHandlerService {
         current = current[path[i] as keyof typeof current];
       }
     }
+  }
+
+  private async updateEndUserWithAmlData({
+    sessionId,
+    endUserId,
+    amlHits,
+    withActiveMonitoring,
+    projectId,
+  }: {
+    sessionId: string;
+    endUserId: string;
+    amlHits: unknown[];
+    withActiveMonitoring: boolean;
+    projectId: TProjectId;
+  }) {
+    const endUser = await this.endUserService.getById(endUserId, {}, [projectId]);
+
+    return await this.endUserService.updateById(endUserId, {
+      data: {
+        amlHits: amlHits as InputJsonValue,
+        ...(withActiveMonitoring
+          ? {
+              activeMonitorings: [
+                ...(endUser.activeMonitorings as z.infer<typeof EndUserActiveMonitoringsSchema>),
+                {
+                  type: 'aml',
+                  vendor: 'veriff',
+                  monitoredUntil: new Date(
+                    new Date().setFullYear(new Date().getFullYear() + 3),
+                  ).toISOString(),
+                  sessionId,
+                },
+              ],
+            }
+          : {}),
+      },
+    });
   }
 }