+++ packages/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/common
 
+## 0.9.6
+
+### Patch Changes
+
+- Context schema update
+
+## 0.9.5
+
+### Patch Changes
+
+- updated enums
+
 ## 0.9.4
 
 ### Patch Changes