+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # web-ui-sdk
 
+## 1.5.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.6
+
+## 1.5.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.5
+
 ## 1.5.5
 
 ### Patch Changes